separate modules in documentable_formatter
commit-id:6ed86016
