+++ crates/cairo-lang-doc/src/lib.rs
@@ -2,6 +2,8 @@ pub mod db;
 pub mod documentable_formatter;
 pub mod documentable_item;
 pub mod parser;
+pub mod signature_data;
+pub mod signature_errors;
 
 #[cfg(test)]
 mod tests;
